Start: Fix issue with --stop when /bin/sh is bash 5.x.
Bash 5.x appears to insert a subshell for executing back-quote
evaluations even when there is only a single binary. This caused
the pid registered for the start-script in _roxen_pid to be for
that subshell under some circumstances. This in turn caused
start --stop not being able to shut down the server (and thus
caused it to behave like --restart).